+++ extensions/sn_remove_dock_glow/change_log.md
@@ -5,3 +5,5 @@ Change Log
   - Initial version.
 * 1.1
   - Fixed content.xml text nodes.
+* 1.2
+  - Added some safety against game version changes.